added entrypoints for deployment of www and dapp resources
